models/site_web_api/espn_site_web_api_client/models/injury_report.py
import datetime
from typing import Any, Dict, List, Type, TypeVar, Union

from attrs import define as _attrs_define
from attrs import field as _attrs_field
from dateutil.parser import isoparse

from ..types import UNSET, Unset

T = TypeVar("T", bound="InjuryReport")


@_attrs_define
class InjuryReport:
    """
    Attributes:
        status (str):  Example: Questionable.
        date (datetime.datetime):  Example: 2023-02-10T18:00:00Z.
        description (Union[Unset, str]):  Example: Ankle.
    """

    status: str
    date: datetime.datetime
    description: Union[Unset, str] = UNSET
    additional_properties: Dict[str, Any] = _attrs_field(init=False, factory=dict)

    def to_dict(self) -> Dict[str, Any]:
        status = self.status

        date = self.date.isoformat()

        description = self.description

        field_dict: Dict[str, Any] = {}
        field_dict.update(self.additional_properties)
        field_dict.update(
            {
                "status": status,
                "date": date,
            }
        )
        if description is not UNSET:
            field_dict["description"] = description

        return field_dict

    @classmethod
    def from_dict(cls: Type[T], src_dict: Dict[str, Any]) -> T:
        d = src_dict.copy()
        status = d.pop("status")

        date = isoparse(d.pop("date"))

        description = d.pop("description", UNSET)

        injury_report = cls(
            status=status,
            date=date,
            description=description,
        )

        injury_report.additional_properties = d
        return injury_report

    @property
    def additional_keys(self) -> List[str]:
        return list(self.additional_properties.keys())

    def __getitem__(self, key: str) -> Any:
        return self.additional_properties[key]

    def __setitem__(self, key: str, value: Any) -> None:
        self.additional_properties[key] = value

    def __delitem__(self, key: str) -> None:
        del self.additional_properties[key]

    def __contains__(self, key: str) -> bool:
        return key in self.additional_properties
